requirements.txt
fastapi
uvicorn
pydantic
groq
psycopg2-binary